feat: add basic grid based movement and attack
Player can select a tile by clicking on it, if tile is occupied, player can select either a Hero or an Enemy. If Player selects the hero and clicks on an unoccupied Tile, hero will relocate on that tile, and if player selects the hero and clicks on a Tile that is occupied by an Enemy, the Enemy will be destroyed.